+++ cli/core/src/typings/index.ts
@@ -1,0 +1,8 @@
+export interface CheckProjectConfiguration {
+  src: string;
+  output: string;
+}
+
+export interface Input {
+  [k: string]: string;
+}