+++ cli/create-naily/src/i18n/utils/manager.ts
@@ -1,10 +1,12 @@
 export interface I18n {
   projectName: string;
   projectNameFileOrFolderExists: string;
+  projectNameTestFailed: string;
   git: string;
   packageManager: string;
   installMyself: string;
   template: string;
+  done: string;
 }
 
 export const languages = new Map<string[], I18n>();